AIAI-395 Create APO full blown real world use case demo - Refactored to dataclasses
