+++ smart-workflow-demo/src/com/axonivy/utils/smart/workflow/demo/erp/supplier/onboarding/enums/AgentStepStatus.java
@@ -1,0 +1,19 @@
+package com.axonivy.utils.smart.workflow.demo.erp.supplier.onboarding.enums;
+
+
+public enum AgentStepStatus {
+  PENDING("Pending"),
+  RUNNING("Running"),
+  COMPLETED("Completed"),
+  FAILED("Failed");
+
+  private final String description;
+
+  AgentStepStatus(String description) {
+    this.description = description;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+}
