+++ smart-workflow-demo/src/com/axonivy/utils/smart/workflow/demo/erp/supplier/onboarding/enums/ApprovalDecision.java
@@ -1,8 +1,6 @@
 package com.axonivy.utils.smart.workflow.demo.erp.supplier.onboarding.enums;
 
-import dev.langchain4j.model.output.structured.Description;
 
-@Description("Decision captured in an approval stage")
 public enum ApprovalDecision {
   APPROVED         ("so-badge-green"),
   REJECTED         ("so-badge-red"),
