+++ smart-workflow-demo/src/com/axonivy/utils/smart/workflow/demo/erp/supplier/onboarding/enums/ApprovalStage.java
@@ -1,8 +1,6 @@
 package com.axonivy.utils.smart.workflow.demo.erp.supplier.onboarding.enums;
 
-import dev.langchain4j.model.output.structured.Description;
 
-@Description("Approval stage in the green happy path")
 public enum ApprovalStage {
   SUPERVISOR,
   QM_ISM
