+++ smart-workflow-demo/src/com/axonivy/utils/smart/workflow/demo/erp/supplier/onboarding/enums/AuditEntryType.java
@@ -1,8 +1,6 @@
 package com.axonivy.utils.smart.workflow.demo.erp.supplier.onboarding.enums;
 
-import dev.langchain4j.model.output.structured.Description;
 
-@Description("Type of audit trail entry — drives which fields and rendering apply")
 public enum AuditEntryType {
   REQUEST_SUBMITTED,
   DUPLICATE_CHECK,
