+++ smart-workflow-demo/src/com/axonivy/utils/smart/workflow/demo/erp/supplier/onboarding/enums/LogLineSeverity.java
@@ -1,0 +1,18 @@
+package com.axonivy.utils.smart.workflow.demo.erp.supplier.onboarding.enums;
+
+
+public enum LogLineSeverity {
+  OK("Ok"),
+  WARNING("Warning"),
+  ERROR("Error");
+
+  private final String description;
+
+  LogLineSeverity(String description) {
+    this.description = description;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+}
